+++ web_plugin/src/main/java/com/futureworkshops/mobileworkflow/plugin/web/WebPlugin.kt
@@ -7,8 +7,8 @@ package com.futureworkshops.mobileworkflow.plugin.web
 import com.futureworkshops.mobileworkflow.domain.DeserializeStep
 import com.futureworkshops.mobileworkflow.plugin.web.step.WebPluginStep
 import com.futureworkshops.mobileworkflow.plugin.web.view.UIWebPluginStep
-import com.futureworkshops.mobileworkflow.surveykit.StepIdentifier
-import com.futureworkshops.mobileworkflow.surveykit.steps.Step
+import com.futureworkshops.mobileworkflow.StepIdentifier
+import com.futureworkshops.mobileworkflow.steps.Step
 
 internal class WebPlugin : DeserializeStep<WebPluginStep>(
     type = "io.mobileworkflow.WebView",
